test: move existing tests to correct files, remove duplicates
