+++ crates/nu-command/tests/commands/format/filesize.rs
@@ -1,3 +1,4 @@
+use nu_test_support::fs::Stub;
 use nu_test_support::prelude::*;
 
 #[test]
@@ -22,4 +23,48 @@ fn format_filesize_with_invalid_unit() -> Result {
     let err = test().run(code).expect_error()?;
     assert!(matches!(err, ShellError::InvalidUnit { .. }));
     Ok(())
+}
+
+#[test]
+fn format_filesize_works() -> Result {
+    Playground::setup("format_filesize_test_1", |dirs, sandbox| {
+        sandbox.with_files(&[
+            Stub::EmptyFile("yehuda.txt"),
+            Stub::EmptyFile("jttxt"),
+            Stub::EmptyFile("andres.txt"),
+        ]);
+
+        let code = "
+            ls
+            | format filesize kB size
+            | get size
+            | first
+        ";
+
+        test().cwd(dirs.test()).run(code).expect_value_eq("0 kB")
+    })
+}
+
+#[test]
+fn format_filesize_works_with_nonempty_files() -> Result {
+    Playground::setup(
+        "format_filesize_works_with_nonempty_files",
+        |dirs, sandbox| {
+            sandbox.with_files(&[Stub::FileWithContentToBeTrimmed(
+                "sample.toml",
+                r#"
+                    [dependency]
+                    name = "nu"
+                "#,
+            )]);
+
+            let code = "ls sample.toml | format filesize B size | get size | first";
+            #[cfg(not(windows))]
+            let expected = "25 B";
+            #[cfg(windows)]
+            let expected = "27 B";
+
+            test().cwd(dirs.test()).run(code).expect_value_eq(expected)
+        },
+    )
 }
